# swagger.yml
openapi: 3.0.0
paths:
...
  /api/departments:
    get:
      operationId: DepartmentsController_list
      parameters:
        - name: isActive
          required: false
          in: query
          description: Filter by active state (true/false)
          schema:
            type: string
        - name: departmentType
          required: false
          in: query
          description: Filter by department type
          schema:
            type: string
        - name: locationId
          required: false
          in: query
          description: Filter by identity site (location) id
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: List departments with search and filters
      tags: &ref_5
        - departments
    post:
      operationId: DepartmentsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDepartmentDto'
      responses:
        '201':
          description: ''
      summary: Create a department (belongs to an identity site location)
      tags: *ref_5
  /api/departments/{id}:
    get:
      operationId: DepartmentsController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get department by id
      tags: *ref_5
    patch:
      operationId: DepartmentsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDepartmentDto'
      responses:
        '200':
          description: ''
      summary: Update a department
      tags: *ref_5
    delete:
      operationId: DepartmentsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Archive a department
      tags: *ref_5
...
  /api/beds:
    get:
      operationId: BedsController_list
      parameters: []
      responses:
        '200':
          description: ''
      summary: List beds with ward/status filters
      tags: &ref_7
        - beds
    post:
      operationId: BedsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBedDto'
      responses:
        '201':
          description: ''
      summary: Create a bed in a ward
      tags: *ref_7
  /api/beds/{id}:
    get:
      operationId: BedsController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get bed by id
      tags: *ref_7
    patch:
      operationId: BedsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBedDto'
      responses:
        '200':
          description: ''
      summary: Update a bed
      tags: *ref_7
    delete:
      operationId: BedsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Archive a bed
      tags: *ref_7
  /api/beds/{id}/status:
    patch:
      operationId: BedsController_setStatus
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Set bed status directly (maintenance/out-of-service)
      tags: *ref_7
...
  /api/encounters:
    get:
      operationId: EncountersController_list
      parameters: []
      responses:
        '200':
          description: ''
      summary: List encounters with patient/visit filters
      tags: &ref_11
        - encounters
    post:
      operationId: EncountersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEncounterDto'
      responses:
        '201':
          description: ''
      summary: Create an encounter (optionally linked to a visit)
      tags: *ref_11
  /api/encounters/{id}:
    get:
      operationId: EncountersController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get encounter by id
      tags: *ref_11
    patch:
      operationId: EncountersController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEncounterDto'
      responses:
        '200':
          description: ''
      summary: Update an encounter
      tags: *ref_11
    delete:
      operationId: EncountersController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Archive an encounter
      tags: *ref_11
  /api/encounters/{id}/requests:
    post:
      operationId: EncountersController_createRequest
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEncounterRequestDto'
      responses:
        '201':
          description: ''
      summary: Create a request from an encounter context
      tags: *ref_11
...
  /api/form-definitions:
    get:
      operationId: FormDefinitionsController_list
      parameters: []
      responses:
        '200':
          description: ''
      summary: List form definitions
      tags: &ref_13
        - form-definitions
    post:
      operationId: FormDefinitionsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFormDefinitionDto'
      responses:
        '201':
          description: ''
      summary: Create a form definition (draft)
      tags: *ref_13
  /api/form-definitions/{id}:
    get:
      operationId: FormDefinitionsController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get form definition by id
      tags: *ref_13
    patch:
      operationId: FormDefinitionsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFormDefinitionDto'
      responses:
        '200':
          description: ''
      summary: Update a form definition (bumps version when schema changes)
      tags: *ref_13
    delete:
      operationId: FormDefinitionsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Archive a form definition
      tags: *ref_13
  /api/form-definitions/{id}/publish:
    post:
      operationId: FormDefinitionsController_publish
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishFormDto'
      responses:
        '201':
          description: ''
      summary: Publish a form definition
      tags: *ref_13
  /api/form-definitions/{id}/unpublish:
    post:
      operationId: FormDefinitionsController_unpublish
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      summary: Unpublish a form definition
      tags: *ref_13
...
info:
  title: EMR API
  description: >-
    Electronic Medical Record: appointments, visits, encounters, dynamic forms,
    and clinical requests
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateDepartmentDto:
      type: object
      properties:
        code:
          type: string
          example: RADIOLOGY
        name:
          type: string
          example: Radiology
        departmentType:
          type: string
          enum: &ref_18
            - OPD
            - INPATIENT
            - EMERGENCY
            - LABORATORY
            - PHARMACY
            - RADIOLOGY
            - MATERNITY
            - SUPPORT
            - OTHER
          default: OPD
        description:
          type: string
        locationId:
          type: string
          description: >-
            Identity location (site) id the department belongs to; defaults to
            the caller’s location
        isActive:
          type: boolean
          default: true
      required:
        - code
        - name
    UpdateDepartmentDto:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        departmentType:
          type: string
          enum: *ref_18
        description:
          type: string
        locationId:
          type: string
          description: Identity location (site) id the department belongs to
        isActive:
          type: boolean
...
    CreateBedDto:
      type: object
      properties:
        code:
          type: string
          description: Short bed code, e.g. B-101
        wardId:
          type: string
          description: Ward the bed belongs to
        bedType:
          type: string
          enum: &ref_20
            - STANDARD
            - ICU
            - MATERNITY
            - ISOLATION
            - PEDIATRIC
            - RECOVERY
          default: STANDARD
        status:
          type: string
          enum: &ref_21
            - AVAILABLE
            - OCCUPIED
            - MAINTENANCE
            - OUT_OF_SERVICE
          default: AVAILABLE
        notes:
          type: string
      required:
        - code
        - wardId
    UpdateBedDto:
      type: object
      properties:
        code:
          type: string
        wardId:
          type: string
        bedType:
          type: string
          enum: *ref_20
        status:
          type: string
          enum: *ref_21
        notes:
          type: string
...
    CreateEncounterDto:
      type: object
      properties:
        patientId:
          type: string
        visitId:
          type: string
        encounterType:
          type: string
          enum: &ref_26
            - CONSULTATION
            - VITALS
            - HISTORY_AND_PHYSICAL
            - CLINICAL_NOTE
            - LAB_RESULTS
            - DISCHARGE
            - PROCEDURE
            - ADMISSION
            - OTHER
        providerId:
          type: string
        providerName:
          type: string
        encounterDatetime:
          type: string
        reason:
          type: string
        notes:
          type: string
      required:
        - patientId
        - encounterType
    CreateEncounterRequestDto:
      type: object
      properties:
        requestType:
          type: string
          enum: &ref_27
            - PRESCRIPTION
            - LAB
            - RADIOLOGY
            - OTHER_TEST
        priority:
          type: string
          enum: *ref_24
          default: ROUTINE
        orderingProviderId:
          type: string
        orderingProviderName:
          type: string
        diagnosis:
          type: string
        clinicalNotes:
          type: string
        requestedAt:
          type: string
        items:
          type: array
          items:
            type: string
      required:
        - requestType
        - items
    UpdateEncounterDto:
      type: object
      properties:
        patientId:
          type: string
        visitId:
          type: string
        encounterType:
          type: string
          enum: *ref_26
        providerId:
          type: string
        providerName:
          type: string
        encounterDatetime:
          type: string
        reason:
          type: string
        notes:
          type: string
...
    CreateFormDefinitionDto:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: &ref_28
            - CLINICAL_NOTE
            - VITALS
            - ASSESSMENT
            - SCREENING
            - PROCEDURE
            - OTHER
        schemaJson:
          type: object
        version:
          type: number
      required:
        - code
        - name
        - category
        - schemaJson
    PublishFormDto:
      type: object
      properties:
        notes:
          type: string
    UpdateFormDefinitionDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: *ref_28
        schemaJson:
          type: object
        version:
          type: number
...